Add conditional rendering on the search button

+++ src/components/SearchBar.js
@@ -12,35 +12,38 @@ export class SearchBar extends Component {
 
     _handleSubmit = async e => {
         e.preventDefault()
-
-        sessionStorage && sessionStorage.clear()
-        this.setState({loading: true})
-
-        if (this.state.inputArtist.length !== 0) {
-            await fetch(`https://ws.audioscrobbler.com/2.0/?method=artist.getinfo&artist=${this.state.inputArtist}&api_key=${process.env.REACT_APP_API_KEY}&format=json`)
-            .then(res => res.json())
-            .then(res => {
-                this.props.onResults(res.artist)
-            })
-            await this.setState({inputArtist: '', loading: false})
+        if (this.state.inputArtist) {
+            
+            sessionStorage && sessionStorage.clear()
+            this.setState({loading: true})
+            
+            if (this.state.inputArtist.length !== 0) {
+                await fetch(`https://ws.audioscrobbler.com/2.0/?method=artist.getinfo&artist=${this.state.inputArtist}&api_key=${process.env.REACT_APP_API_KEY}&format=json`)
+                .then(res => res.json())
+                .then(res => {
+                    this.props.onResults(res.artist)
+                })
+                await this.setState({inputArtist: '', loading: false})
+            }
         }
     }
 
     render() {
+        
         return (
             <form onSubmit={this._handleSubmit}>
             <div className="field has-addons">
                 <div className={this.state.loading ? "control is-loading" : "control"} >
                     <input
                     onChange={this._handleChange}
-                    className="input" 
+                    className="input is-radiusless" 
                     type="text" 
                     placeholder="Search for an Artist"
                     value={this.state.inputArtist}
                     />
                 </div>
                 <div className="control">
-                    <button className="button is-warning">
+                    <button className={`button is-warning is-radiusless ${this.state.inputArtist.length > 0 ? '' : 'is-hidden'}`}>
                         <p className="has-text-weight-bold">SEARCH</p>
                     </button>
                 </div>
